Set version as 1.0.0.0
TODO - add in type to the JSON so you cant connect to a DCS SRS server

+++ DCS-SR-Client/Network/IL2/Models/ClientDataMessage.cs
@@ -8,16 +8,16 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Client.Network.IL2.Models
 {
     public class ClientDataMessage : IL2UDPMessage
     {
-        public long ClientID { get; set; }
-        public long ParentVehicleClientID { get; set; }
-        public long ServerClientID { get; set; }
+        public int ClientID { get; set; }
+        public int ParentVehicleClientID { get; set; }
+        public int ServerClientID { get; set; }
         public short Coalition { get; set; }
 
         public ClientDataMessage(byte[] message, int offset)
         {
-            ClientID = BitConverter.ToInt64(message, offset);
-            ParentVehicleClientID = BitConverter.ToInt64(message, offset+ 4);
-            ServerClientID = BitConverter.ToInt64(message, offset + 4 +4);
+            ClientID = BitConverter.ToInt32(message, offset);
+            ParentVehicleClientID = BitConverter.ToInt32(message, offset+ 4);
+            ServerClientID = BitConverter.ToInt32(message, offset + 4 +4);
             Coalition = BitConverter.ToInt16(message, offset+ 4+4+4);
         }
